:recycle: Refactor IDex to remove dependency from Orca implementation
