+++ src/Solana.Unity.Dex/Orca/Quotes/IncreaseLiquidityQuoteUtils.cs
@@ -1,3 +1,4 @@
+using Solana.Unity.Dex.Math;
 using System;
 using System.Numerics;
 
@@ -7,6 +8,7 @@ using Solana.Unity.Dex.Orca.Math;
 using Solana.Unity.Dex.Orca.Core.Accounts;
 using Solana.Unity.Dex.Orca.Ticks;
 using Solana.Unity.Dex.Orca.Positions;
+using Solana.Unity.Dex.Quotes;
 
 namespace Solana.Unity.Dex.Orca.Quotes
 {
@@ -207,14 +209,5 @@ namespace Solana.Unity.Dex.Orca.Quotes
         public int TickLowerIndex { get; set; }
         public int TickUpperIndex { get; set; }
         public Percentage SlippageTolerance { get; set; }
-    }
-
-    public class IncreaseLiquidityQuote
-    {
-        public BigInteger LiquidityAmount { get; set; }
-        public BigInteger TokenMaxA { get; set; }
-        public BigInteger TokenMaxB { get; set; }
-        public BigInteger TokenEstA { get; set; }
-        public BigInteger TokenEstB { get; set; }
     }
 }
