+++ src/Solana.Unity.Dex/Orca/Swap/PoolUtils.cs
@@ -1,3 +1,4 @@
+using Solana.Unity.Dex.Math;
 using System;
 using System.Numerics;
 using System.Collections.Generic;
@@ -8,6 +9,7 @@ using Solana.Unity.Dex.Orca.Math;
 using Solana.Unity.Dex.Orca.Address;
 using Solana.Unity.Dex.Orca.Core.Types;
 using Solana.Unity.Dex.Orca.Core.Accounts;
+using Solana.Unity.Dex.Swap;
 using BigDecimal = Solana.Unity.Dex.Orca.Math.BigDecimal;
 
 namespace Solana.Unity.Dex.Orca.Swap
