+++ src/Solana.Unity.Dex/Orca/Swap/SwapUtils.cs
@@ -1,3 +1,4 @@
+using Solana.Unity.Dex.Math;
 using System;
 using System.Linq;
 using System.Numerics;
@@ -11,6 +12,7 @@ using Solana.Unity.Dex.Orca.Math;
 using Solana.Unity.Dex.Orca.Ticks;
 using Solana.Unity.Dex.Orca.Core.Accounts;
 using Solana.Unity.Dex.Orca.Address;
+using Solana.Unity.Dex.Swap;
 
 namespace Solana.Unity.Dex.Orca.Swap
 {
