--- src/Solana.Unity.Dex/Orca/Swap/TokenType.cs
@@ -1,12 +1,0 @@
-namespace Solana.Unity.Dex.Orca.Swap
-{
-    /// <summary>
-    /// An enum for the token type in a Whirlpool.
-    /// </summary>
-    public enum TokenType
-    {
-        TokenA = 1,
-        TokenB,
-        None
-    }
-}
