+++ src/Solana.Unity.Dex/Orca/TxApi/Dex.cs
@@ -1,3 +1,4 @@
+using Solana.Unity.Dex.Math;
 using System;
 using System.Numerics; 
 using System.Threading.Tasks;
@@ -6,13 +7,10 @@ using Solana.Unity.Rpc;
 using Solana.Unity.Rpc.Models;
 using Solana.Unity.Wallet;
 using Solana.Unity.Rpc.Types;
-
-using Solana.Unity.Dex.Orca.Swap;
-using Solana.Unity.Dex.Orca.Math;
-using Solana.Unity.Dex.Orca.Quotes;
-using Solana.Unity.Dex.Orca.Quotes.Swap;
 using Solana.Unity.Dex.Orca.Core;
 using Solana.Unity.Dex.Orca.Core.Accounts;
+using Solana.Unity.Dex.Quotes;
+using Solana.Unity.Dex.Swap;
 
 namespace Solana.Unity.Dex.Orca.TxApi
 {
@@ -38,7 +36,7 @@ namespace Solana.Unity.Dex.Orca.TxApi
         public abstract Task<Transaction> Swap(
             PublicKey whirlpoolAddress,
             ulong amount,
-            bool aToB,
+            bool aToB = true,
             PublicKey tokenAuthority = null,
             Commitment commitment = Commitment.Finalized
         );
@@ -159,7 +157,16 @@ namespace Solana.Unity.Dex.Orca.TxApi
             Commitment commitment
         );
 
-        /// <inheritdoc />
+        /// <summary>
+        /// Attempts to find an existing whirlpool with the specified properties, using address derivation 
+        /// and then checking for the existence of such an account. 
+        /// </summary> 
+        /// <param name="tokenMintA">Mint address of any token associated with the pool, preferably token A.</param> 
+        /// <param name="tokenMintB">Mint address of any token associated with the pool, preferably token B.</param> 
+        /// <param name="tickSpacing">Preferred tickSpacing associated with the pool; if not found, others will be queried.</param> 
+        /// <param name="configAccountAddress">Public key of the whirlpool config address account.</param>
+        /// <param name="commitment">Transaction commitment to use for chain queries.</param> 
+        /// <returns>A tuple of the whirlpool's address, and the Whirlpool instance.</returns>
         public abstract Task<Tuple<PublicKey, Whirlpool>> FindWhirlpool(
             PublicKey tokenMintA,
             PublicKey tokenMintB,
