+++ src/Solana.Unity.Dex/Swap/TokenType.cs
@@ -1,0 +1,11 @@
+namespace Solana.Unity.Dex.Swap;
+
+/// <summary>
+/// An enum for the token type in a Whirlpool.
+/// </summary>
+public enum TokenType
+{
+    TokenA = 1,
+    TokenB,
+    None
+}
