+++ test/Solana.Unity.Dex.Test/Orca/Integration/InitializePoolTests.cs
@@ -8,6 +8,7 @@ using Solana.Unity.Dex.Orca.Address;
 using Solana.Unity.Dex.Orca.Core.Accounts;
 using Solana.Unity.Dex.Test.Orca.Params;
 using Solana.Unity.Dex.Test.Orca.Utils;
+using Solana.Unity.Dex.Ticks;
 using Solana.Unity.Rpc.Core.Http;
 using Solana.Unity.Rpc.Types;
