+++ test/Solana.Unity.Dex.Test/Orca/Integration/TxApi/FindWhirlpoolTests.cs
@@ -20,6 +20,7 @@ using Solana.Unity.Dex.Orca.Core.Accounts;
 using Solana.Unity.Dex.Test.Orca.Params;
 using Solana.Unity.Dex.Orca.Address;
 using Solana.Unity.Dex.Test.Orca.Utils;
+using Solana.Unity.Dex.Ticks;
 
 namespace Solana.Unity.Dex.Test.Orca.Integration.TxApi
 {
