+++ test/Solana.Unity.Dex.Test/Orca/SdkTests/UtilsTests/PoolUtilsTests.cs
@@ -9,6 +9,7 @@ using Solana.Unity.Dex.Orca.Swap;
 using Solana.Unity.Dex.Orca.Address;
 using Solana.Unity.Dex.Test.Orca.Utils;
 using Solana.Unity.Dex.Orca.Core.Accounts;
+using Solana.Unity.Dex.Swap;
 
 namespace Solana.Unity.Dex.Test.Orca.Sdk.Utils
 {
