+++ test/Solana.Unity.Dex.Test/Orca/Utils/FeeTierTestUtils.cs
@@ -9,6 +9,7 @@ using Solana.Unity.Dex.Orca.Ticks;
 using Solana.Unity.Dex.Orca.Address;
 using Solana.Unity.Dex.Test.Orca.Params;
 using Solana.Unity.Dex.Orca.Core.Program;
+using Solana.Unity.Dex.Ticks;
 using Solana.Unity.Rpc.Models;
 
 namespace Solana.Unity.Dex.Test.Orca.Utils
